requirements.txt
selenium==4.24.0
webdriver-manager==4.0.2
pytest==8.3.3
allure-pytest==2.13.5
Faker==29.0.0
python-dotenv==1.0.1